refactor enum class
